Add error rollback, remove dead code, extract shared model
Add error handlers to ignore/unignore calls that revert local state on backend failure. Remove unused BehaviorSubject from PreferencesService. Extract MediaLibrary and ActiveServerResponse into shared media-server model so non-Plex components no longer import from plex.model.
