+++ frontend/src/app/models/media-server.model.ts
@@ -1,0 +1,10 @@
+export interface MediaLibrary {
+  title: string;
+  type: string;  // 'movie', 'show', 'artist', 'photo'
+}
+
+export interface ActiveServerResponse {
+  server: string;
+  token: string;
+  libraries: MediaLibrary[];
+}
