+++ frontend/src/app/models/plex.model.ts
@@ -1,3 +1,5 @@
+export { MediaLibrary as PlexLibrary, ActiveServerResponse } from './media-server.model';
+
 export interface PlexAuthResponse {
   oauth_url: string;
 }
@@ -5,15 +7,4 @@ export interface PlexAuthResponse {
 export interface PlexServersResponse {
   servers: string[];
   token: string;
-}
-
-export interface PlexLibrary {
-  title: string;
-  type: string;  // 'movie', 'show', 'artist', 'photo'
-}
-
-export interface ActiveServerResponse {
-  server: string;
-  token: string;
-  libraries: PlexLibrary[];
 }
