+++ frontend/src/app/services/preferences.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable, BehaviorSubject } from 'rxjs';
-import { tap } from 'rxjs/operators';
+import { Observable } from 'rxjs';
 import { environment } from '../../environments/environment';
 
 export interface UserPreferences {
@@ -17,24 +16,14 @@ export interface UserPreferences {
   providedIn: 'root'
 })
 export class PreferencesService {
-  private prefsSubject = new BehaviorSubject<UserPreferences | null>(null);
-  prefs$ = this.prefsSubject.asObservable();
 
   constructor(private http: HttpClient) {}
 
   load(): Observable<UserPreferences> {
-    return this.http.get<UserPreferences>(`${environment.apiUrl}/preferences`).pipe(
-      tap(prefs => this.prefsSubject.next(prefs))
-    );
+    return this.http.get<UserPreferences>(`${environment.apiUrl}/preferences`);
   }
 
   save(prefs: Partial<UserPreferences>): Observable<UserPreferences> {
-    return this.http.post<UserPreferences>(`${environment.apiUrl}/preferences`, prefs).pipe(
-      tap(saved => this.prefsSubject.next(saved))
-    );
-  }
-
-  get current(): UserPreferences | null {
-    return this.prefsSubject.value;
+    return this.http.post<UserPreferences>(`${environment.apiUrl}/preferences`, prefs);
   }
 }
